:recycle: Refactor
Add propstypes
