+++ src/pages/Missions/Missions.css
@@ -48,46 +48,13 @@
   border: 1px solid rgb(129, 126, 126);
 }
 
-.li-content:nth-child(even) {
-  background-color: rgba(148, 147, 147, 0.502);
-}
-
 .table-container ul li {
   border: 1px solid rgb(129, 126, 126);
   display: flex;
 }
 
-.name {
-  font-size: 20px;
-  font-weight: bold;
-  border: none;
-  flex: 1.5;
-  margin: 0;
-}
-
-.des {
-  flex: 6;
-}
-
-.content {
-  border-left: 1px solid rgb(129, 126, 126);
-  height: 350px;
-  padding-left: 5px;
-}
-
-.status {
-  flex: 1.5;
-}
-
 .join {
   flex: 2;
-}
-
-.btn1 {
-  background-color: rgb(129, 126, 126);
-  font-size: 11px;
-  color: #fff;
-  padding: 6px;
 }
 
 .btn2 {
@@ -99,8 +66,3 @@
   border: 1px solid red;
 }
 
-.active-btn {
-  background-color: rgb(43, 175, 236);
-  color: #fff;
-  padding: 6px;
-}
